+++ qt/python/mantidqt/widgets/plotconfigdialog/curvestabwidget/__init__.py
@@ -14,7 +14,7 @@ from matplotlib.lines import Line2D
 from numpy import isclose
 from qtpy.QtCore import Qt
 
-from mantid.plots import MantidAxes
+from mantid.plots.mantidaxes import MantidAxes
 from mantid.plots.datafunctions import errorbars_hidden
 from mantidqt.widgets.plotconfigdialog.colorselector import convert_color_to_hex
